prototype to neq
